+++ compiler-core/src/language_server/engine.rs
@@ -13,7 +13,7 @@ use crate::{
     line_numbers::LineNumbers,
     paths::ProjectPaths,
     type_::{
-        self, pretty::Printer, Deprecation, ModuleInterface, Type, TypeConstructor,
+        self, printer::Printer, Deprecation, ModuleInterface, Type, TypeConstructor,
         ValueConstructorVariant,
     },
     Error, Result, Warning,
@@ -366,7 +366,9 @@ where
                         symbols.push(DocumentSymbol {
                             name: name.to_string(),
                             detail: Some(
-                                Printer::new().pretty_print(&get_function_type(function), 0),
+                                Printer::new(&module.ast.names)
+                                    .print_type(&get_function_type(function))
+                                    .to_string(),
                             ),
                             kind: SymbolKind::FUNCTION,
                             tags: make_deprecated_symbol_tag(&function.deprecation),
@@ -392,7 +394,14 @@ where
                         #[allow(deprecated)]
                         symbols.push(DocumentSymbol {
                             name: alias.alias.to_string(),
-                            detail: Some(Printer::new().pretty_print(&alias.type_, 0)),
+                            detail: Some(
+                                Printer::new(&module.ast.names)
+                                    // If we print with aliases, we end up printing the alias which the user
+                                    // is currently hovering, which is not helpful. Instead, we print the
+                                    // raw type, so the user can see which type the alias represents
+                                    .print_type_without_aliases(&alias.type_)
+                                    .to_string(),
+                            ),
                             kind: SymbolKind::CLASS,
                             tags: make_deprecated_symbol_tag(&alias.deprecation),
                             deprecated: None,
@@ -406,7 +415,7 @@ where
                     }
 
                     Definition::CustomType(type_) => {
-                        symbols.push(custom_type_symbol(type_, &line_numbers));
+                        symbols.push(custom_type_symbol(type_, &line_numbers, module));
                     }
 
                     Definition::ModuleConstant(constant) => {
@@ -431,7 +440,11 @@ where
                         #[allow(deprecated)]
                         symbols.push(DocumentSymbol {
                             name: constant.name.to_string(),
-                            detail: Some(Printer::new().pretty_print(&constant.type_, 0)),
+                            detail: Some(
+                                Printer::new(&module.ast.names)
+                                    .print_type(&constant.type_)
+                                    .to_string(),
+                            ),
                             kind: SymbolKind::CONSTANT,
                             tags: make_deprecated_symbol_tag(&constant.deprecation),
                             deprecated: None,
@@ -499,40 +512,50 @@ where
                 None => return Ok(None),
             };
 
+            let Some(module) = this.module_for_uri(&params.text_document.uri) else {
+                return Ok(None);
+            };
+
             Ok(match found {
                 Located::Statement(_) => None, // TODO: hover for statement
                 Located::ModuleStatement(Definition::Function(fun)) => {
-                    Some(hover_for_function_head(fun, lines))
+                    Some(hover_for_function_head(fun, lines, module))
                 }
                 Located::ModuleStatement(Definition::ModuleConstant(constant)) => {
-                    Some(hover_for_module_constant(constant, lines))
+                    Some(hover_for_module_constant(constant, lines, module))
                 }
                 Located::ModuleStatement(_) => None,
                 Located::UnqualifiedImport(UnqualifiedImport {
                     name,
-                    module,
+                    module: module_name,
                     is_type,
                     location,
                 }) => this
                     .compiler
-                    .get_module_interface(module.as_str())
-                    .and_then(|module| {
+                    .get_module_interface(module_name.as_str())
+                    .and_then(|module_interface| {
                         if is_type {
-                            module.types.get(name).map(|t| {
-                                hover_for_annotation(*location, t.type_.as_ref(), Some(t), lines)
+                            module_interface.types.get(name).map(|t| {
+                                hover_for_annotation(
+                                    *location,
+                                    t.type_.as_ref(),
+                                    Some(t),
+                                    lines,
+                                    module,
+                                )
                             })
                         } else {
-                            module.values.get(name).map(|v| {
-                                let m = if this.hex_deps.contains(&module.package) {
-                                    Some(module)
+                            module_interface.values.get(name).map(|v| {
+                                let m = if this.hex_deps.contains(&module_interface.package) {
+                                    Some(module_interface)
                                 } else {
                                     None
                                 };
-                                hover_for_imported_value(v, location, lines, m, name)
+                                hover_for_imported_value(v, location, lines, m, name, module)
                             })
                         }
                     }),
-                Located::Pattern(pattern) => Some(hover_for_pattern(pattern, lines)),
+                Located::Pattern(pattern) => Some(hover_for_pattern(pattern, lines, module)),
                 Located::PatternSpread {
                     spread_location,
                     arguments,
@@ -549,10 +572,11 @@ where
                             continue;
                         }
 
-                        let type_ = Printer::new().pretty_print(argument.value.type_().as_ref(), 0);
+                        let type_ = Printer::new(&module.ast.names)
+                            .print_type(argument.value.type_().as_ref());
                         match &argument.label {
-                            Some(label) => labelled.push(format!("- `{}: {}`", label, type_)),
-                            None => positional.push(format!("- `{}`", type_)),
+                            Some(label) => labelled.push(format!("- `{label}: {type_}`")),
+                            None => positional.push(format!("- `{type_}`")),
                         }
                     }
 
@@ -575,17 +599,13 @@ Unused labelled fields:
                         range,
                     })
                 }
-                Located::Expression(expression) => {
-                    let module = this.module_for_uri(&params.text_document.uri);
-
-                    Some(hover_for_expression(
-                        expression,
-                        lines,
-                        module,
-                        &this.hex_deps,
-                    ))
-                }
-                Located::Arg(arg) => Some(hover_for_function_argument(arg, lines)),
+                Located::Expression(expression) => Some(hover_for_expression(
+                    expression,
+                    lines,
+                    module,
+                    &this.hex_deps,
+                )),
+                Located::Arg(arg) => Some(hover_for_function_argument(arg, lines, module)),
                 Located::FunctionBody(_) => None,
                 Located::Annotation(annotation, type_) => {
                     let type_constructor = type_constructor_from_modules(
@@ -597,6 +617,7 @@ Unused labelled fields:
                         &type_,
                         type_constructor,
                         lines,
+                        module,
                     ))
                 }
             })
@@ -661,7 +682,11 @@ Unused labelled fields:
     }
 }
 
-fn custom_type_symbol(type_: &CustomType<Arc<Type>>, line_numbers: &LineNumbers) -> DocumentSymbol {
+fn custom_type_symbol(
+    type_: &CustomType<Arc<Type>>,
+    line_numbers: &LineNumbers,
+    module: &Module,
+) -> DocumentSymbol {
     let constructors = type_
         .constructors
         .iter()
@@ -688,7 +713,11 @@ fn custom_type_symbol(type_: &CustomType<Arc<Type>>, line_numbers: &LineNumbers)
                 #[allow(deprecated)]
                 arguments.push(DocumentSymbol {
                     name: label.to_string(),
-                    detail: Some(Printer::new().pretty_print(&argument.type_, 0)),
+                    detail: Some(
+                        Printer::new(&module.ast.names)
+                            .print_type(&argument.type_)
+                            .to_string(),
+                    ),
                     kind: SymbolKind::FIELD,
                     tags: None,
                     deprecated: None,
@@ -770,11 +799,11 @@ fn custom_type_symbol(type_: &CustomType<Arc<Type>>, line_numbers: &LineNumbers)
     }
 }
 
-fn hover_for_pattern(pattern: &TypedPattern, line_numbers: LineNumbers) -> Hover {
+fn hover_for_pattern(pattern: &TypedPattern, line_numbers: LineNumbers, module: &Module) -> Hover {
     let documentation = pattern.get_documentation().unwrap_or_default();
 
     // Show the type of the hovered node to the user
-    let type_ = Printer::new().pretty_print(pattern.type_().as_ref(), 0);
+    let type_ = Printer::new(&module.ast.names).print_type(pattern.type_().as_ref());
     let contents = format!(
         "```gleam
 {type_}
@@ -794,7 +823,11 @@ fn get_function_type(fun: &TypedFunction) -> Type {
     }
 }
 
-fn hover_for_function_head(fun: &TypedFunction, line_numbers: LineNumbers) -> Hover {
+fn hover_for_function_head(
+    fun: &TypedFunction,
+    line_numbers: LineNumbers,
+    module: &Module,
+) -> Hover {
     let empty_str = EcoString::from("");
     let documentation = fun
         .documentation
@@ -802,7 +835,7 @@ fn hover_for_function_head(fun: &TypedFunction, line_numbers: LineNumbers) -> Ho
         .map(|(_, doc)| doc)
         .unwrap_or(&empty_str);
     let function_type = get_function_type(fun);
-    let formatted_type = Printer::new().pretty_print(&function_type, 0);
+    let formatted_type = Printer::new(&module.ast.names).print_type(&function_type);
     let contents = format!(
         "```gleam
 {formatted_type}
@@ -815,8 +848,12 @@ fn hover_for_function_head(fun: &TypedFunction, line_numbers: LineNumbers) -> Ho
     }
 }
 
-fn hover_for_function_argument(argument: &TypedArg, line_numbers: LineNumbers) -> Hover {
-    let type_ = Printer::new().pretty_print(&argument.type_, 0);
+fn hover_for_function_argument(
+    argument: &TypedArg,
+    line_numbers: LineNumbers,
+    module: &Module,
+) -> Hover {
+    let type_ = Printer::new(&module.ast.names).print_type(&argument.type_);
     let contents = format!("```gleam\n{type_}\n```");
     Hover {
         contents: HoverContents::Scalar(MarkedString::String(contents)),
@@ -829,12 +866,17 @@ fn hover_for_annotation(
     annotation_type: &Type,
     type_constructor: Option<&TypeConstructor>,
     line_numbers: LineNumbers,
+    module: &Module,
 ) -> Hover {
     let empty_str = EcoString::from("");
     let documentation = type_constructor
         .and_then(|t| t.documentation.as_ref())
         .unwrap_or(&empty_str);
-    let type_ = Printer::new().pretty_print(annotation_type, 0);
+    // If a user is hovering an annotation, it's not very useful to show the
+    // local representation of that type, since that's probably what they see
+    // in the source code anyway. So here, we print the raw type,
+    // which is probably more helpful.
+    let type_ = Printer::new(&module.ast.names).print_type_without_aliases(annotation_type);
     let contents = format!(
         "```gleam
 {type_}
@@ -850,9 +892,10 @@ fn hover_for_annotation(
 fn hover_for_module_constant(
     constant: &ModuleConstant<Arc<Type>, EcoString>,
     line_numbers: LineNumbers,
+    module: &Module,
 ) -> Hover {
     let empty_str = EcoString::from("");
-    let type_ = Printer::new().pretty_print(&constant.type_, 0);
+    let type_ = Printer::new(&module.ast.names).print_type(&constant.type_);
     let documentation = constant
         .documentation
         .as_ref()
@@ -868,20 +911,19 @@ fn hover_for_module_constant(
 fn hover_for_expression(
     expression: &TypedExpr,
     line_numbers: LineNumbers,
-    module: Option<&Module>,
+    module: &Module,
     hex_deps: &std::collections::HashSet<EcoString>,
 ) -> Hover {
     let documentation = expression.get_documentation().unwrap_or_default();
 
-    let link_section = module
-        .and_then(|m: &Module| {
-            let (module_name, name) = get_expr_qualified_name(expression)?;
-            get_hexdocs_link_section(module_name, name, &m.ast, hex_deps)
+    let link_section = get_expr_qualified_name(expression)
+        .and_then(|(module_name, name)| {
+            get_hexdocs_link_section(module_name, name, &module.ast, hex_deps)
         })
         .unwrap_or("".to_string());
 
     // Show the type of the hovered node to the user
-    let type_ = Printer::new().pretty_print(expression.type_().as_ref(), 0);
+    let type_ = Printer::new(&module.ast.names).print_type(expression.type_().as_ref());
     let contents = format!(
         "```gleam
 {type_}
@@ -900,6 +942,7 @@ fn hover_for_imported_value(
     line_numbers: LineNumbers,
     hex_module_imported_from: Option<&ModuleInterface>,
     name: &EcoString,
+    module: &Module,
 ) -> Hover {
     let documentation = value.get_documentation().unwrap_or_default();
 
@@ -908,7 +951,7 @@ fn hover_for_imported_value(
     });
 
     // Show the type of the hovered node to the user
-    let type_ = Printer::new().pretty_print(value.type_.as_ref(), 0);
+    let type_ = Printer::new(&module.ast.names).print_type(value.type_.as_ref());
     let contents = format!(
         "```gleam
 {type_}
@@ -1112,7 +1155,7 @@ fn code_action_fix_names(
                 new_text: correction.to_string(),
             };
 
-            CodeActionBuilder::new(&format!("Rename to {}", correction))
+            CodeActionBuilder::new(&format!("Rename to {correction}"))
                 .kind(lsp_types::CodeActionKind::QUICKFIX)
                 .changes(uri.clone(), vec![edit])
                 .preferred(true)