+++ compiler-core/src/type_/printer.rs
@@ -1,19 +1,15 @@
-#![allow(dead_code)]
-
+use bimap::BiMap;
 use ecow::EcoString;
-use im::{HashMap, HashSet};
-use std::sync::Arc;
+use im::HashMap;
+use std::{collections::HashSet, sync::Arc};
 
 use crate::type_::{Type, TypeVar};
 
 /// This class keeps track of what names are used for modules in the current
 /// scope, so they can be printed in errors, etc.
 ///
-#[derive(Debug)]
-pub struct TypeNames {
-    uid: u64,
-    current_module: EcoString,
-
+#[derive(Debug, Clone, PartialEq, Eq, Default)]
+pub struct Names {
     /// Types that exist in the current module, either defined or imported in an
     /// unqualified fashion.
     ///
@@ -47,7 +43,7 @@ pub struct TypeNames {
     /// - key:   `("some/module", "Wibble")`
     /// - value: `"Wobble"`
     ///
-    local_types: HashMap<(EcoString, EcoString), EcoString>,
+    local_types: BiMap<(EcoString, EcoString), EcoString>,
 
     /// Mapping of imported modules to their locally used named
     ///
@@ -92,18 +88,66 @@ pub struct TypeNames {
     /// value: `"something"`
     ///
     type_variables: HashMap<u64, EcoString>,
-    type_variable_names: HashSet<EcoString>,
+
+    /// Constructors which are imported in the current module in an
+    /// unqualified fashion.
+    ///
+    /// key:   (Defining module name, type name)
+    /// value: Alias name
+    ///
+    /// # Example 1
+    ///
+    /// ```gleam
+    /// import wibble.{Wobble}
+    /// ```
+    /// would result in
+    /// - key:   `("wibble", "Wobble")`
+    /// - value: `"Wobble"`
+    ///
+    /// # Example 2
+    ///
+    /// ```gleam
+    /// import wibble.{Wobble as Woo}
+    /// ```
+    /// would result in
+    /// - key:   `("wibble", "Wobble")`
+    /// - value: `"Woo"`
+    ///
+    local_value_constructors: HashMap<(EcoString, EcoString), EcoString>,
+
+    /// A map from local constructor names to the modules which they refer to.
+    /// This helps resolve cases like:
+    /// ```gleam
+    /// import wibble.{Wobble}
+    /// type Wibble { Wobble }
+    /// ```
+    /// Here, `Wobble` is shadowed, causing `Wobble` not to be valid syntax
+    /// for `wibble.Wobble`.
+    ///
+    /// Each key is the local name of the constructor, and the value is the module
+    /// for which the unqualified version is valid. In the above example,
+    /// it would result in
+    /// - key:   `"Wobble"`
+    /// - value: `"module"` (Whatever the current module is)
+    ///
+    /// But in this case:
+    /// ```gleam
+    /// import wibble.{Wobble as Wubble}
+    /// type Wibble { Wobble }
+    /// ```
+    /// No shadowing occurs, so this isn't needed.
+    ///
+    constructor_names: HashMap<EcoString, EcoString>,
 }
 
-impl TypeNames {
-    pub fn new(current_module: EcoString) -> Self {
+impl Names {
+    pub fn new() -> Self {
         Self {
-            uid: Default::default(),
-            current_module,
             local_types: Default::default(),
             imported_modules: Default::default(),
             type_variables: Default::default(),
-            type_variable_names: Default::default(),
+            local_value_constructors: Default::default(),
+            constructor_names: Default::default(),
         }
     }
 
@@ -114,15 +158,26 @@ impl TypeNames {
         type_name: EcoString,
         local_alias: EcoString,
     ) {
+        _ = self.local_types.remove_by_right(&local_alias);
         _ = self
             .local_types
             .insert((module_name, type_name), local_alias);
     }
 
+    pub fn type_in_scope(&mut self, local_alias: EcoString, type_: &Type) {
+        match type_ {
+            Type::Named { module, name, .. } => {
+                self.named_type_in_scope(module.clone(), name.clone(), local_alias);
+            }
+            Type::Fn { .. } | Type::Var { .. } | Type::Tuple { .. } => {
+                _ = self.local_types.remove_by_right(&local_alias);
+            }
+        }
+    }
+
     /// Record a type variable in this module.
     pub fn type_variable_in_scope(&mut self, id: u64, local_alias: EcoString) {
         _ = self.type_variables.insert(id, local_alias.clone());
-        _ = self.type_variable_names.insert(local_alias);
     }
 
     /// Record an imported module in this module.
@@ -135,33 +190,231 @@ impl TypeNames {
         &'a self,
         module: &'a EcoString,
         name: &'a EcoString,
-    ) -> NamedTypeNames<'a> {
+        print_mode: PrintMode,
+    ) -> NameContextInformation<'a> {
+        if print_mode == PrintMode::ExpandAliases {
+            if let Some(module) = self.imported_modules.get(module) {
+                return NameContextInformation::Qualified(module, name.as_str());
+            };
+
+            return NameContextInformation::Unimported(name.as_str());
+        }
+
         let key = (module.clone(), name.clone());
 
+        // Only check for local aliases if we want to print aliases
         // There is a local name for this type, use that.
-        if let Some(name) = self.local_types.get(&key) {
-            return NamedTypeNames::Unqualified(name.as_str());
+        if let Some(name) = self.local_types.get_by_left(&key) {
+            return NameContextInformation::Unqualified(name.as_str());
         }
 
         // This type is from a module that has been imported
         if let Some(module) = self.imported_modules.get(module) {
-            return NamedTypeNames::Qualified(module, name.as_str());
+            return NameContextInformation::Qualified(module, name.as_str());
         };
 
-        return NamedTypeNames::Unimported(name.as_str());
+        return NameContextInformation::Unimported(name.as_str());
+    }
+
+    /// Record a named value in this module.
+    pub fn named_constructor_in_scope(
+        &mut self,
+        module_name: EcoString,
+        value_name: EcoString,
+        local_alias: EcoString,
+    ) {
+        _ = self
+            .local_value_constructors
+            .insert((module_name.clone(), value_name), local_alias.clone());
+        _ = self.constructor_names.insert(local_alias, module_name);
+    }
+
+    /// Get the name and optional module qualifier for a named constructor.
+    pub fn named_constructor<'a>(
+        &'a self,
+        module: &'a EcoString,
+        name: &'a EcoString,
+    ) -> NameContextInformation<'a> {
+        let key: (EcoString, EcoString) = (module.clone(), name.clone());
+
+        // There is a local name for this value, use that.
+        if let Some(name) = self.local_value_constructors.get(&key) {
+            // Only return unqualified syntax if the constructor is not shadowed,
+            // and unqualified syntax is valid.
+            if self
+                .constructor_names
+                .get(name)
+                .expect("Constructors must be added to both maps")
+                == module
+            {
+                return NameContextInformation::Unqualified(name.as_str());
+            }
+        }
+
+        // This value is from a module that has been imported
+        if let Some(module) = self.imported_modules.get(module) {
+            return NameContextInformation::Qualified(module, name.as_str());
+        };
+
+        NameContextInformation::Unimported(name.as_str())
+    }
+}
+
+#[derive(Debug)]
+pub enum NameContextInformation<'a> {
+    /// This type is from a module that has not been imported in this module.
+    Unimported(&'a str),
+    /// This type has been imported in an unqualifid fashion in this module.
+    Unqualified(&'a str),
+    /// This type is from a module that has been imported.
+    Qualified(&'a str, &'a str),
+}
+
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum PrintMode {
+    /// Prints the context-specific representation of a type.
+    Normal,
+    /// Prints full detail of the given type, always qualified.
+    /// Useful for providing more detail to the user.
+    ///
+    /// For example, with this code:
+    /// ```gleam
+    /// type A = Int
+    /// ```
+    /// If the type `gleam.Int` were printed using the `Normal` mode,
+    /// we would print `A`, since that is the local alias for the `Int` type.
+    ///
+    /// However, if the user were hovering over the type `A` itself, it wouldn't be
+    /// particularly helpful to print `A`.
+    /// So with `ExpandAliases`, it would print `gleam.Int`,
+    /// which tells the user exactly what type `A` represents.
+    ///
+    ExpandAliases,
+}
+
+/// A type printer that does not wrap and indent, but does take into account the
+/// names that types and modules have been aliased with in the current module.
+#[derive(Debug)]
+pub struct Printer<'a> {
+    names: &'a Names,
+    uid: u64,
+
+    /// Some type variables aren't bound to names, so when trying to print those,
+    /// we need to create our own names which don't overlap with existing type variables.
+    /// These two data structures store a mapping of IDs to created type-variable names,
+    /// to ensure consistent printing, and the set of all printed names so that we don't
+    /// create a type variable name which matches an existing one.
+    ///
+    /// Note: These are stored per printer, not per TypeNames struct, because:
+    /// - It doesn't really matter what these are, as long as they are consistent.
+    /// - We would need mutable access to the names struct, which isn't really possible
+    ///   in many contexts.
+    ///
+    printed_type_variables: HashMap<u64, EcoString>,
+    printed_type_variable_names: HashSet<EcoString>,
+}
+
+impl<'a> Printer<'a> {
+    pub fn new(names: &'a Names) -> Self {
+        Printer {
+            names,
+            uid: Default::default(),
+            printed_type_variables: Default::default(),
+            printed_type_variable_names: names.type_variables.values().cloned().collect(),
+        }
+    }
+
+    pub fn print_type(&mut self, type_: &Type) -> EcoString {
+        let mut buffer = EcoString::new();
+        self.print(type_, &mut buffer, PrintMode::Normal);
+        buffer
+    }
+
+    pub fn print_type_without_aliases(&mut self, type_: &Type) -> EcoString {
+        let mut buffer = EcoString::new();
+        self.print(type_, &mut buffer, PrintMode::ExpandAliases);
+        buffer
+    }
+
+    fn print(&mut self, type_: &Type, buffer: &mut EcoString, print_mode: PrintMode) {
+        match type_ {
+            Type::Named {
+                name, args, module, ..
+            } => {
+                let (module, name) = match self.names.named_type(module, name, print_mode) {
+                    NameContextInformation::Qualified(m, n) => (Some(m), n),
+                    NameContextInformation::Unqualified(n) => (None, n),
+                    // TODO: indicate that the module is not import and as such
+                    // needs to be, as well as how.
+                    NameContextInformation::Unimported(n) => {
+                        (Some(module.split('/').last().unwrap_or(module)), n)
+                    }
+                };
+
+                if let Some(module) = module {
+                    buffer.push_str(module);
+                    buffer.push('.');
+                }
+                buffer.push_str(name);
+
+                if !args.is_empty() {
+                    buffer.push('(');
+                    self.print_arguments(args, buffer, print_mode);
+                    buffer.push(')');
+                }
+            }
+
+            Type::Fn { args, retrn } => {
+                buffer.push_str("fn(");
+                self.print_arguments(args, buffer, print_mode);
+                buffer.push_str(") -> ");
+                self.print(retrn, buffer, print_mode);
+            }
+
+            Type::Var { type_, .. } => match *type_.borrow() {
+                TypeVar::Link { ref type_, .. } => self.print(type_, buffer, print_mode),
+                TypeVar::Unbound { id, .. } | TypeVar::Generic { id, .. } => {
+                    buffer.push_str(&self.type_variable(id))
+                }
+            },
+
+            Type::Tuple { elems, .. } => {
+                buffer.push_str("#(");
+                self.print_arguments(elems, buffer, print_mode);
+                buffer.push(')');
+            }
+        }
+    }
+
+    fn print_arguments(
+        &mut self,
+        args: &[Arc<Type>],
+        typ_str: &mut EcoString,
+        print_mode: PrintMode,
+    ) {
+        for (i, arg) in args.iter().enumerate() {
+            self.print(arg, typ_str, print_mode);
+            if i < args.len() - 1 {
+                typ_str.push_str(", ");
+            }
+        }
     }
 
     /// A suitable name of a type variable.
     pub fn type_variable(&mut self, id: u64) -> EcoString {
-        if let Some(name) = self.type_variables.get(&id) {
+        if let Some(name) = self.names.type_variables.get(&id) {
+            return name.clone();
+        }
+
+        if let Some(name) = self.printed_type_variables.get(&id) {
             return name.clone();
         }
 
         loop {
             let name = self.next_letter();
-            if !self.type_variable_names.contains(&name) {
-                _ = self.type_variable_names.insert(name.clone());
-                _ = self.type_variables.insert(id, name.clone());
+            if !self.printed_type_variable_names.contains(&name) {
+                _ = self.printed_type_variable_names.insert(name.clone());
+                _ = self.printed_type_variables.insert(id, name.clone());
                 return name;
             }
         }
@@ -190,99 +443,11 @@ impl TypeNames {
     }
 }
 
-#[derive(Debug)]
-pub enum NamedTypeNames<'a> {
-    /// This type is from a module that has not been imported in this module.
-    Unimported(&'a str),
-    /// This type has been imported in an unqualifid fashion in this module.
-    Unqualified(&'a str),
-    /// This type is from a module that has been imported.
-    Qualified(&'a str, &'a str),
-}
-
-/// A type printer that does not wrap and indent, but does take into account the
-/// names that types and modules have been aliased with in the current module.
-#[derive(Debug)]
-pub struct Printer<'a> {
-    names: &'a mut TypeNames,
-}
-
-impl<'a> Printer<'a> {
-    pub fn new(names: &'a mut TypeNames) -> Self {
-        Printer { names }
-    }
-
-    pub fn print_type(&mut self, type_: &Type) -> EcoString {
-        let mut buffer = EcoString::new();
-        self.print(type_, &mut buffer);
-        buffer
-    }
-
-    fn print(&mut self, type_: &Type, buffer: &mut EcoString) {
-        match type_ {
-            Type::Named {
-                name, args, module, ..
-            } => {
-                let (module, name) = match self.names.named_type(module, name) {
-                    NamedTypeNames::Qualified(m, n) => (Some(m), n),
-                    NamedTypeNames::Unqualified(n) => (None, n),
-                    // TODO: indicate that the module is not import and as such
-                    // needs to be, as well as how.
-                    NamedTypeNames::Unimported(n) => {
-                        (Some(module.split('/').last().unwrap_or(module)), n)
-                    }
-                };
-
-                if let Some(module) = module {
-                    buffer.push_str(module);
-                    buffer.push('.');
-                }
-                buffer.push_str(name);
-
-                if !args.is_empty() {
-                    buffer.push('(');
-                    self.print_arguments(args, buffer);
-                    buffer.push(')');
-                }
-            }
-
-            Type::Fn { args, retrn } => {
-                buffer.push_str("fn(");
-                self.print_arguments(args, buffer);
-                buffer.push_str(") -> ");
-                self.print(retrn, buffer);
-            }
-
-            Type::Var { type_, .. } => match *type_.borrow() {
-                TypeVar::Link { ref type_, .. } => self.print(type_, buffer),
-                TypeVar::Unbound { id, .. } | TypeVar::Generic { id, .. } => {
-                    buffer.push_str(&self.names.type_variable(id))
-                }
-            },
-
-            Type::Tuple { elems, .. } => {
-                buffer.push_str("#(");
-                self.print_arguments(elems, buffer);
-                buffer.push(')');
-            }
-        }
-    }
-
-    fn print_arguments(&mut self, args: &[Arc<Type>], typ_str: &mut EcoString) {
-        for (i, arg) in args.iter().enumerate() {
-            self.print(arg, typ_str);
-            if i < args.len() - 1 {
-                typ_str.push_str(", ");
-            }
-        }
-    }
-}
-
 #[test]
 fn test_local_type() {
-    let mut names = TypeNames::new("module".into());
+    let mut names = Names::new();
     names.named_type_in_scope("mod".into(), "Tiger".into(), "Cat".into());
-    let mut printer = Printer::new(&mut names);
+    let mut printer = Printer::new(&names);
 
     let type_ = Type::Named {
         name: "Tiger".into(),
@@ -296,10 +461,47 @@ fn test_local_type() {
 }
 
 #[test]
+fn test_prelude_type() {
+    let mut names = Names::new();
+    names.named_type_in_scope("gleam".into(), "Int".into(), "Int".into());
+    let mut printer = Printer::new(&names);
+
+    let type_ = Type::Named {
+        name: "Int".into(),
+        args: vec![],
+        module: "gleam".into(),
+        publicity: crate::ast::Publicity::Public,
+        package: "".into(),
+    };
+
+    assert_eq!(printer.print_type(&type_), "Int");
+}
+
+#[test]
+fn test_shadowed_prelude_type() {
+    let mut names = Names::new();
+
+    names.named_type_in_scope("gleam".into(), "Int".into(), "Int".into());
+    names.named_type_in_scope("mod".into(), "Int".into(), "Int".into());
+
+    let mut printer = Printer::new(&names);
+
+    let type_ = Type::Named {
+        name: "Int".into(),
+        args: vec![],
+        module: "gleam".into(),
+        publicity: crate::ast::Publicity::Public,
+        package: "".into(),
+    };
+
+    assert_eq!(printer.print_type(&type_), "gleam.Int");
+}
+
+#[test]
 fn test_generic_type_annotation() {
-    let mut names = TypeNames::new("module".into());
+    let mut names = Names::new();
     names.type_variable_in_scope(0, "one".into());
-    let mut printer = Printer::new(&mut names);
+    let mut printer = Printer::new(&names);
 
     let type_ = Type::Var {
         type_: Arc::new(std::cell::RefCell::new(TypeVar::Generic { id: 0 })),
@@ -310,8 +512,8 @@ fn test_generic_type_annotation() {
 
 #[test]
 fn test_generic_type_var() {
-    let mut names = TypeNames::new("module".into());
-    let mut printer = Printer::new(&mut names);
+    let names = Names::new();
+    let mut printer = Printer::new(&names);
 
     let type_ = Type::Var {
         type_: Arc::new(std::cell::RefCell::new(TypeVar::Unbound { id: 0 })),
@@ -327,8 +529,8 @@ fn test_generic_type_var() {
 
 #[test]
 fn test_tuple_type() {
-    let mut names = TypeNames::new("module".into());
-    let mut printer = Printer::new(&mut names);
+    let names = Names::new();
+    let mut printer = Printer::new(&names);
 
     let type_ = Type::Tuple {
         elems: vec![
@@ -354,10 +556,10 @@ fn test_tuple_type() {
 
 #[test]
 fn test_fn_type() {
-    let mut names = TypeNames::new("module".into());
+    let mut names = Names::new();
     names.named_type_in_scope("gleam".into(), "Int".into(), "Int".into());
     names.named_type_in_scope("gleam".into(), "Bool".into(), "Bool".into());
-    let mut printer = Printer::new(&mut names);
+    let mut printer = Printer::new(&names);
 
     let type_ = Type::Fn {
         args: vec![
@@ -390,9 +592,9 @@ fn test_fn_type() {
 
 #[test]
 fn test_module_alias() {
-    let mut names = TypeNames::new("module".into());
+    let mut names = Names::new();
     names.imported_module("mod1".into(), "animals".into());
-    let mut printer = Printer::new(&mut names);
+    let mut printer = Printer::new(&names);
 
     let type_ = Type::Named {
         name: "Cat".into(),
@@ -407,13 +609,13 @@ fn test_module_alias() {
 
 #[test]
 fn test_type_alias_and_generics() {
-    let mut names = TypeNames::new("module".into());
+    let mut names = Names::new();
 
     names.named_type_in_scope("mod".into(), "Tiger".into(), "Cat".into());
 
     names.type_variable_in_scope(0, "one".into());
 
-    let mut printer = Printer::new(&mut names);
+    let mut printer = Printer::new(&names);
 
     let type_ = Type::Named {
         name: "Tiger".into(),
@@ -430,13 +632,13 @@ fn test_type_alias_and_generics() {
 
 #[test]
 fn test_unqualified_import_and_generic() {
-    let mut names = TypeNames::new("module".into());
+    let mut names = Names::new();
 
     names.named_type_in_scope("mod".into(), "Cat".into(), "C".into());
 
     names.type_variable_in_scope(0, "one".into());
 
-    let mut printer = Printer::new(&mut names);
+    let mut printer = Printer::new(&names);
 
     let type_ = Type::Named {
         name: "Cat".into(),
@@ -453,8 +655,8 @@ fn test_unqualified_import_and_generic() {
 
 #[test]
 fn nested_module() {
-    let mut names = TypeNames::new("module".into());
-    let mut printer = Printer::new(&mut names);
+    let names = Names::new();
+    let mut printer = Printer::new(&names);
     let type_ = Type::Named {
         name: "Cat".into(),
         args: vec![],
@@ -468,7 +670,7 @@ fn nested_module() {
 
 #[test]
 fn test_unqualified_import_and_module_alias() {
-    let mut names = TypeNames::new("module".into());
+    let mut names = Names::new();
 
     names.imported_module("mod1".into(), "animals".into());
 
@@ -476,7 +678,7 @@ fn test_unqualified_import_and_module_alias() {
         .local_types
         .insert(("mod1".into(), "Cat".into()), "C".into());
 
-    let mut printer = Printer::new(&mut names);
+    let mut printer = Printer::new(&names);
 
     let type_ = Type::Named {
         name: "Cat".into(),
@@ -491,13 +693,13 @@ fn test_unqualified_import_and_module_alias() {
 
 #[test]
 fn test_module_imports() {
-    let mut names = TypeNames::new("module".into());
+    let mut names = Names::new();
     names.imported_module("mod".into(), "animals".into());
     let _ = names
         .local_types
         .insert(("mod2".into(), "Cat".into()), "Cat".into());
 
-    let mut printer = Printer::new(&mut names);
+    let mut printer = Printer::new(&names);
 
     let type_ = Type::Named {
         name: "Cat".into(),
@@ -521,12 +723,12 @@ fn test_module_imports() {
 
 #[test]
 fn test_multiple_generic_annotations() {
-    let mut names = TypeNames::new("module".into());
+    let mut names = Names::new();
 
     names.type_variable_in_scope(0, "one".into());
     names.type_variable_in_scope(1, "two".into());
 
-    let mut printer = Printer::new(&mut names);
+    let mut printer = Printer::new(&names);
 
     let type_ = Type::Named {
         name: "Tiger".into(),
@@ -553,12 +755,12 @@ fn test_multiple_generic_annotations() {
 
 #[test]
 fn test_variable_name_already_in_scope() {
-    let mut names = TypeNames::new("module".into());
+    let mut names = Names::new();
 
     names.type_variable_in_scope(1, "a".into());
     names.type_variable_in_scope(2, "b".into());
 
-    let mut printer = Printer::new(&mut names);
+    let mut printer = Printer::new(&names);
 
     let type_ = |id| Type::Var {
         type_: Arc::new(std::cell::RefCell::new(TypeVar::Generic { id })),