+++ compiler-wasm/src/lib.rs
@@ -78,7 +78,7 @@ fn get_warnings(project_id: usize) -> VectorWarningEmitterIO {
 #[wasm_bindgen]
 pub fn write_module(project_id: usize, module_name: &str, code: &str) {
     let fs = get_filesystem(project_id);
-    let path = format!("/src/{}.gleam", module_name);
+    let path = format!("/src/{module_name}.gleam");
     fs.write(&Utf8PathBuf::from(path), code)
         .expect("writing file")
 }
@@ -132,7 +132,7 @@ pub fn compile_package(project_id: usize, target: &str) -> Result<(), String> {
 #[wasm_bindgen]
 pub fn read_compiled_javascript(project_id: usize, module_name: &str) -> Option<String> {
     let fs = get_filesystem(project_id);
-    let path = format!("/build/{}.mjs", module_name);
+    let path = format!("/build/{module_name}.mjs");
     fs.read(&Utf8PathBuf::from(path)).ok()
 }